backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
torch
opencv-python
openpyxl

# Local models (installed via editable path or sys.path)
# sam3
# RoMa
